separated concerns for string utilities.Other concerns are inside the utilities function but for now it's enough. close #8
